+++ addon/controllers/front.js
@@ -1,8 +1,8 @@
 import Ember from 'ember';
-import GetController from 'ember-pretenderify/controllers/get';
-import PostController from 'ember-pretenderify/controllers/post';
-import PutController from 'ember-pretenderify/controllers/put';
-import DeleteController from 'ember-pretenderify/controllers/delete';
+import GetController from 'ember-cli-mirage/controllers/get';
+import PostController from 'ember-cli-mirage/controllers/post';
+import PutController from 'ember-cli-mirage/controllers/put';
+import DeleteController from 'ember-cli-mirage/controllers/delete';
 
 export default {
 
@@ -24,7 +24,7 @@ export default {
     var handlerMethod = handlerType + 'Handler';
 
     var handlerExists = this[controller][handlerMethod];
-    if (!handlerExists) { console.error('Pretenderify: You passed a ' + handlerType + ' as a handler to the ' + controller + ' but no ' + handlerMethod + ' was implemented.'); return;}
+    if (!handlerExists) { console.error('Mirage: You passed a ' + handlerType + ' as a handler to the ' + controller + ' but no ' + handlerMethod + ' was implemented.'); return;}
 
     var data = this[controller][handlerMethod](handler, store, request, code);