+++ app/initializers/ember-cli-mirage.js
@@ -1,13 +1,13 @@
 import ENV from '../config/environment';
 import userConfig from '../pretender/config';
-import Server from 'ember-pretenderify/server';
-import readData from 'ember-pretenderify/utils/read-data';
-import readFactories from 'ember-pretenderify/utils/read-factories';
+import Server from 'ember-cli-mirage/server';
+import readData from 'ember-cli-mirage/utils/read-data';
+import readFactories from 'ember-cli-mirage/utils/read-factories';
 
 export default {
-  name: 'ember-pretenderify',
+  name: 'ember-cli-mirage',
   initialize: function(container, application) {
-    var config = ENV['ember-pretenderify'];
+    var config = ENV['ember-cli-mirage'];
     var env = ENV.environment;
     var usingInDev = env === 'development' && !config.usingProxy;
     var usingInTest = env === 'test';