+++ tests/fixtures/config/environment-with-force-true.js
@@ -17,7 +17,7 @@ module.exports = function(environment) {
       // Here you can pass flags/options to your application instance
       // when it is created
     },
-    'ember-pretenderify': {
+    'ember-cli-mirage': {
       force: true
     }
   };